+++ src/execution/ExecutionTask.ts
@@ -3,6 +3,7 @@ import { BehaviorSubject, concat, from } from 'rxjs';
 import { PartialDeep } from 'type-fest';
 import { string_SCREAMING_CASE } from '../_packages/utils.index';
 import type { task_id } from '../types/typeAliases';
+import { TODO_USE } from '../utils/organization/TODO_USE';
 import { $randomToken } from '../utils/random/$randomToken';
 import type { PipelineExecutorResult } from './PipelineExecutorResult';
 
@@ -46,7 +47,11 @@ export async function createTask<TTask extends AbstractTask<TTaskResult>, TTaskR
     return {
         taskType,
         taskId,
-        asPromise() {
+        asPromise(options) {
+            const { isCrashedOnError = true } = options || {};
+
+            TODO_USE(isCrashedOnError);
+
             return /* not await */ finalResult;
         },
         asObservable() {
@@ -65,7 +70,7 @@ export type ExecutionTask = AbstractTask<PipelineExecutorResult> & {
 
 /**
  * Represents a task that prepares a pipeline
- * @deprecated Currently unused
+ * @deprecated TODO: [🐚] Currently unused - use
  */
 export type PreparationTask = AbstractTask<PipelineExecutorResult> & {
     readonly taskType: 'PREPARATION';
@@ -89,7 +94,7 @@ export type AbstractTask<TTaskResult> = {
     /**
      * Gets a promise that resolves with the task result
      */
-    asPromise(): Promise<TTaskResult>;
+    asPromise(options?: { readonly isCrashedOnError?: boolean }): Promise<TTaskResult>;
 
     /**
      * Gets an observable stream of partial task results