+++ src/execution/PipelineExecutor.ts
@@ -1,7 +1,7 @@
 import type { Promisable } from 'type-fest';
 import type { TaskProgress } from '../types/TaskProgress';
 import type { InputParameters } from '../types/typeAliases';
-import type { PipelineExecutorResult } from './PipelineExecutorResult';
+import { ExecutionTask } from './ExecutionTask';
 
 /**
  * Executor is a simple async function that takes INPUT  PARAMETERs and returns result parameters _(along with all intermediate parameters and INPUT  PARAMETERs = it extends input object)_.
@@ -15,10 +15,10 @@ export type PipelineExecutor = {
     (
         inputParameters: InputParameters,
         onProgress?: (taskProgress: TaskProgress) => Promisable<void>,
-    ): Promise<PipelineExecutorResult>;
+    ): Promise<ExecutionTask>;
 };
 
 /**
- * TODO: [🐚] Change onProgress to object that represents the running execution, can be subscribed via RxJS to and also awaited
+ * TODO: [🐚] Maybe return just `ExecutionTask` instead of `Promise<ExecutionTask>`
  * TODO: [🧠] Should this file be in /execution or /types folder?
  */